server code are refactore in folder based
